refactor: 스네이크 표기법 사용된 패키지명 변경
